When running under docker or lxc, don't detect parent process death
There's probably a better solution, but I don't know it. When running
under docker, getppid returns 0 or 1 (I'm not sure which), which is
the same parent id that you get when running outside of a container,
and your parent process died (because init becomes your new parent
process).
Anyway, this seems to be OK for now, because if docker kills a
container, it reaps the child processes too.