Improve parent process exit detection on linux/docker
Previously, when running in a container, we gave up trying
to detect parent process death. Now, instead of using ppid
equal to 0 or 1 as a sign that the parent has died, we simply
check to see if the parent id has changed. We take any change
in parent id as a sign that it has died.