Make logging functions 'const'
I explain the details of this decision in the comment block above
the Logger class.

Basically, the 'const' is not true in the sense that the logging
functions produce side effects. However, most applications would
consider these side-effects to be beyond the scope of the application
that they're building. This is especially true in the case of
uberlog, because of it's multi-process, non-blocking architecture.

And in case it's not obvious, the motivation for making the logging
functions 'const', is so that the Logger object can be used when
it's a member of a larger class/struct, which is held via a
const reference.